Add support for GitHub OAuth authentication
After some experimenting with dedicated github/flask OAuth2 modules, I end up
using Requests-OAuthlib mainly because it is better maintained, stable,
proven and with good documentation.

Perhasp it does not hide all the OAuth2 details like, for instance, flask-oauth,
but what it does is clear and understandable, moreover I am not a big fan
of 'magic' code: when it fails I want to understand why.