Handle missing auth token from GitHub
This can only happen if a spurious GET call is done at /github_callback

I hit this one when micgrating to Openshift, I didin't update the
call back url on github, so upon startinga registration on openshift,
the c9 instance was called instead.

It is more a deploy setup issue than an app error, but check it anyhow.