Update tests results
When a test finishes, Fishtest sends an email with the result to a
Google group that forwards it to its subscribers. Among them there is
an email address belonging to a webhook service that, upon receiving the
mail, forwards it here through a POST request. A bit complex but the
bottom line is we receive a timely update as soon as a test finishes and
we use this info to update the corresponding row in TestsDB.